Fix Windows build failure

+++ headers/simdfastpfor.h
@@ -379,7 +379,7 @@ public:
         datatobepacked(PageSize),
         bytescontainer(PageSize + 3 * PageSize / BlockSize) {
     assert(ps / BlockSize * BlockSize == ps);
-    assert(gccbits(BlockSizeInUnitsOfPackSize * PACKSIZE - 1) <= 8);
+    assert(gccbits(static_cast<uint32_t>(BlockSizeInUnitsOfPackSize * PACKSIZE - 1)) <= 8);
   }
   enum {
     BlockSizeInUnitsOfPackSize = 4,
